refactor(config): split provider catalog into provider_catalog.rs (closes #1082)
Spun out from #1074 (architecture audit, P2 follow-up).

\`koda-core/src/config.rs\` was 1,343 LOC and mixed two unrelated
concerns: static provider catalog (\`ProviderMeta\` + \`ProviderType\`
+ metadata lookup table + URL/name auto-detection) vs. runtime
settings loading (\`ModelSettings\`, \`AgentConfig\`, \`KodaConfig\`,
file I/O, env merging, validation). The audit (#1074 finding #6)
explicitly called this out as the only file in the codebase that
was "not cohesive — splittable".

## Split

- New \`koda-core/src/provider_catalog.rs\` (273 LOC) holds the
  static, compile-time catalog. Pure data — no I/O, no env-var
  reads, no file-system access. Comes with a module doc explaining
  the four-step process for adding a new provider (add variant +
  add \`meta()\` arm + optional name alias + optional URL
  fingerprint + snapshot test row).
- \`koda-core/src/config.rs\` shrinks from 1,343 → 1,109 LOC
  (−234 lines, 17% smaller). Now scoped to runtime settings
  loading exclusively.
- \`pub use crate::provider_catalog::{ProviderMeta, ProviderType}\`
  re-exported in \`config.rs\` so every existing caller —
  \`koda_core::config::ProviderType\` is used in 8 files across
  koda-core / koda-cli / koda-test-utils plus the snapshot test
  — continues to work without a churn-the-codebase import update.
  New code may import directly from \`koda_core::provider_catalog\`
  if preferred.
- \`pub mod provider_catalog;\` registered in \`lib.rs\`.

## Net change

| File | Before | After | Δ |
|---|---|---|---|
| config.rs | 1,343 | 1,109 | −234 |
| provider_catalog.rs | (new) | 273 | +273 |
| lib.rs | (n) | n+1 | +1 |
| **Total** | **1,343** | **1,383** | **+40** |

The 40 extra lines are the new module doc (~30) and the re-export
comment block (~10) — explanation, not duplicate code. Zero
duplication. Zero behavioural change. \`config.rs\` is now off the
"not cohesive" watchlist.

## Acceptance (per #1082)

- ✅ \`wc -l koda-core/src/config.rs\` returns 1,109 (estimate was
  <1,100; off by 9 from the doc/comment additions, well within
  spirit of the criterion)
- ✅ \`wc -l koda-core/src/provider_catalog.rs\` returns 273 (~250)
- ✅ \`cargo test --workspace\` — 1,871 lib + ~440 integration tests
  pass, including \`provider_metadata_all\` snapshot test which
  pins every catalog entry
- ✅ \`cargo clippy --workspace --all-targets -- -D warnings\` clean
- ✅ \`cargo fmt --all\` no diff
- ✅ \`pub use\` re-exports preserve every external import path

## Non-goals respected

- No API changes
- No behavioural changes
- No new tests (existing snapshot test already pins catalog content)
- Did not touch the other 11 large files (audit verdict: cohesive,
  leave alone)

Closes #1082.
